Updated Interface for Different counters
It's time to work on our second counter type, this is the Container which is from where we're going to spawn Kitchen Objects and give them to the player. Again we're going to face the same problem where some of our code is only made to work with a specific ClearCounter and not any Counter so we're going to have to refactor that and this time instead of using an Interface we're going to use Inheritance.